Script that computes the Feedback gains of the LQR controller
The script will compute and write into a txt file the feedback gains of the LQR coontroller. For that Q is fixed at diag(0.004,...,0.004), then R varie between 100 and 10000 with a step of 100.
Moreover,  as the matrix A depends of the reaction wheel speed, the feddback gain is therefore also dependent of the RW speed; which is why the feedback gains that are writen in the txt file are the mean values computed by varying the RW speed between -6000 to 6000 RPM.

NB : In the txt file, each line correspond to a fixed parameter
-1st value : Q
-2nd value : R
-3rd value : kqe1
-4th value : kqe2
-5th value : kqe3
-6th value : kw1
-7th value : kw2
-8th value : kw3